telegram_bot/requirements.txt
python-telegram-bot==21.0
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
aiohttp==3.9.1
httpx==0.27.2
